Clean up create organisation page
